--- src/Validators/AbstractValidator.php
@@ -1,22 +1,0 @@
-<?php
-
-/**
- * JBZoo Toolbox - Csv-Blueprint.
- *
- * This file is part of the JBZoo Toolbox project.
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * @license    MIT
- * @copyright  Copyright (C) JBZoo.com, All rights reserved.
- * @see        https://github.com/JBZoo/Csv-Blueprint
- */
-
-declare(strict_types=1);
-
-namespace JBZoo\CsvBlueprint\Validators;
-
-abstract class AbstractValidator
-{
-    abstract public function validate(?string $cellValue, int $line): ErrorSuite;
-}
